% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{compute_concentrations}
\alias{compute_concentrations}
\title{Compute count conversions from net and pump data}
\usage{
compute_concentrations(df, cols = NULL, aggregate = TRUE)
}
\arguments{
\item{df}{forcis net dataset}

\item{cols}{character, if NULL only required_columns will be included
If TRUE subsample counts will be aggregated by sample_id.}

\item{aggregate}{logical, if FALSE counts will be derived for each subsample.}
}
\value{
dataframe
}
\description{
Compute count conversions from net and pump data
}
