% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_forcis_db.R
\name{download_forcis_db}
\alias{download_forcis_db}
\title{Download the entire FORCIS database}
\usage{
download_forcis_db(
  path = ".",
  version = options()$forcis_version,
  check_for_update = options()$check_for_update,
  overwrite = FALSE,
  timeout = 60
)
}
\arguments{
\item{path}{a \code{character} of length 1. The folder in which the FORCIS
database will be saved. Note that a subdirectory will be created, e.g.
\verb{forcis-db/version-99/} (with \code{99} the version number).}

\item{version}{a \code{character} of length 1. The version number (with two
numbers, e.g. \code{08} instead of \code{8}) of the FORCIS database to download.
Default is the latest version. Note that this argument can be handle with
the global option \code{forcis_version}. For example, if user calls
\code{options(forcis_version = "07")}, the version \code{07} will be used by default
for the current R session. It is recommended to use the latest version of
the database.}

\item{check_for_update}{a \code{logical}. If \code{TRUE} (default) the function will
check if a newer version of the FORCIS database is available on Zenodo and
ask user to download it. Note that this argument can be handle with
the global option \code{check_for_update}. For example, if user calls
\code{options(check_for_update = FALSE)}, invitation to download the latest
version will be disable for the current R session.}

\item{overwrite}{a \code{logical}. If \code{TRUE} it will override the downloaded
files of the FORCIS database. Default is \code{FALSE}.}

\item{timeout}{a \code{integer}. The timeout for downloading files from the
FORCIS database. Default is \code{60}. This number can be increased for low
Internet connection.}
}
\value{
No return value. The FORCIS files will be saved in the \code{path} folder.
}
\description{
Downloads the entire FORCIS database as a collection of five \code{csv} files from
Zenodo (\url{https://zenodo.org/doi/10.5281/zenodo.7390791}). Additional
files are also downloaded (see
\url{https://zenodo.org/doi/10.5281/zenodo.7390791}).
}
\examples{
\dontrun{
# Folder in which the database will be saved ----
path_to_save_db <- "data"

# Download the database ----
download_forcis_db(path = path_to_save_db)

# Check the content of the folder ----
list.files(path_to_save_db, recursive = TRUE)
}
}
