% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.R
\name{extract_species_names}
\alias{extract_species_names}
\title{Extract species names from columns names}
\usage{
extract_species_names(data)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{get_*_data()} functions.}
}
\value{
A \code{data.frame}.
}
\description{
\strong{ADD DESCRIPTION}
}
\examples{
## __ADD EXAMPLE__
}
