% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_coordinates}
\alias{filter_by_coordinates}
\title{Filter forcis data by coordinate square}
\usage{
filter_by_coordinates(data, coord_square)
}
\arguments{
\item{data}{forcis data in long format}

\item{coord_square}{a numeric vector containing in this order minimum latitute,
minimum longitude, maximum latitude, maximum longitude}
}
\value{
A \code{data.frame}.
}
\description{
Filter forcis data by coordinate square
}
