% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcis-package.R
\docType{package}
\name{forcis-package}
\alias{forcis}
\alias{forcis-package}
\title{forcis: Handle the FORCIS Foraminifera Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to the 'FORCIS' database (Chaabane et al. (2024) \doi{10.5281/zenodo.7390791}) on global foraminifera distribution. This package allows to download and to handle 'FORCIS' data. It is part of the FRB-CESAB working group FORCIS. \url{https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/forcis/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/forcis/}
  \item \url{https://github.com/ropensci/forcis}
  \item Report bugs at \url{https://github.com/ropensci/forcis/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]
  \item Mattia Greco \email{mattia_greco@outlook.com} (\href{https://orcid.org/0000-0003-2416-6235}{ORCID})
  \item Sonia Chaabane \email{sonia.chaabane@gmail.com} (\href{https://orcid.org/0000-0002-4653-8610}{ORCID})
  \item Xavier Giraud \email{giraud@cerege.fr} (\href{https://orcid.org/0000-0001-5067-8176}{ORCID})
  \item Thibault de Garidel-Thoron \email{garidel@cerege.fr} (\href{https://orcid.org/0000-0001-8983-9571}{ORCID})
}

Other contributors:
\itemize{
  \item Khalil Hammami \email{khalil.hammami@enetcom.usf.tn} [contributor]
  \item Air Forbes \email{amforbes@ualberta.ca} (\href{https://orcid.org/0000-0002-9842-7648}{ORCID}) [reviewer]
  \item FRB-CESAB [funder]
}

}
\keyword{internal}
