% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_version.R
\name{get_current_version}
\alias{get_current_version}
\title{Get the version of the database currently used}
\usage{
get_current_version()
}
\value{
A \code{character} of length 1.
}
\description{
This function returns the version of the database currently used in the
project.
}
\examples{
## ADD EXAMPLE ----
}
