% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_data}
\alias{get_data}
\alias{get_plankton_nets_data}
\alias{get_pump_data}
\alias{get_cpr_north_data}
\alias{get_cpr_south_data}
\alias{get_sediment_trap_data}
\title{Import FORCIS data}
\usage{
get_plankton_nets_data(
  path = ".",
  version = forcis_db_version(),
  overwrite = FALSE,
  timeout = 60
)

get_pump_data(
  path = ".",
  version = forcis_db_version(),
  overwrite = FALSE,
  timeout = 60
)

get_cpr_north_data(
  path = ".",
  version = forcis_db_version(),
  overwrite = FALSE,
  timeout = 60
)

get_cpr_south_data(
  path = ".",
  version = forcis_db_version(),
  overwrite = FALSE,
  timeout = 60
)

get_sediment_trap_data(
  path = ".",
  version = forcis_db_version(),
  overwrite = FALSE,
  timeout = 60
)
}
\arguments{
\item{path}{a \code{character} of length 1. The folder in which the FORCIS
database has been (or will be) saved.}

\item{version}{a \code{character} of length 1. The version number of the
FORCIS database. Default is the latest version.}

\item{overwrite}{a \code{logical}. If \code{TRUE} it will override the downloaded
files of the FORCIS database. Default is \code{FALSE}.}

\item{timeout}{a \code{integer}. The timeout for downloading files from the
FORCIS database. Default is \code{60}.}
}
\value{
A \code{data.frame}. See \url{https://zenodo.org/record/7390792} for a
preview of the dataset.
}
\description{
These functions import one specific \code{csv} file of the FORCIS database
(see below) stored in the folder \code{path}. If the \code{csv} file does not exist
it will be downloaded from Zenodo (\url{https://zenodo.org/record/7936568}).
}
\details{
\itemize{
\item \code{get_plankton_nets_data()} imports the FORCIS plankton nets data
\item \code{get_pump_data()} imports the FORCIS pump data
\item \code{get_cpr_north_data()} imports the FORCIS CPR North data
\item \code{get_cpr_south_data()} imports the FORCIS CPR South data
\item \code{get_sediment_trap_data()} imports the FORCIS sediment traps data
}
}
\seealso{
\code{\link[=get_forcis_db]{get_forcis_db()}} to download the complete FORCIS database.
}
