% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocean_names.R
\name{get_ocean_names}
\alias{get_ocean_names}
\title{Get World ocean names}
\usage{
get_ocean_names()
}
\value{
A \code{character} vector with World ocean names.
}
\description{
This function returns the name of World oceans according to the IHO Sea Areas
dataset version 3 (Flanders Marine Institute, 2018).
}
\examples{
# Print the name of World oceans ----
get_ocean_names()
}
\references{
Flanders Marine Institute (2018). IHO Sea Areas, version 3.
Available online at: \url{https://www.marineregions.org/}.
DOI: \url{https://doi.org/10.14284/323}.
}
