% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_required_columns.R
\name{get_required_columns}
\alias{get_required_columns}
\title{Get required column names}
\usage{
get_required_columns()
}
\value{
A \code{character}.
}
\description{
Gets required column names (except taxa names) for the package. This
function is designed to help users to add additional columns in
\code{\link[=select_columns]{select_columns()}} (argument \code{cols}) if missing from this list.
}
\examples{
\dontrun{
# Get required column names (expect taxa names) ----
forcis::get_required_columns()
}
}
