% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshape and simplify FORCIS data}
\usage{
reshape_data(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset, except for CPR North data.}
}
\value{
A \code{data.frame} reshaped in a long format.
}
\description{
Reshapes FORCIS data by pivoting species columns into two columns: \code{taxa}
(taxon names) and \code{counts} (taxon abundances). It converts wider \code{data.frame}
to a long format.
}
\examples{
# Attach the package ----
library("forcis")

# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
                         package = "forcis")

net_data <- read.table(file_name, dec = ".", sep = ";")

# Add 'data_type' column ----
net_data$"data_type" <- "Net"

# Dimensions of the data.frame ----
dim(net_data)

# Reshape data ----
net_data <- reshape_data(net_data)

# Dimensions of the data.frame ----
dim(net_data)

# Column names ----
colnames(net_data)
}
