% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_forcis.R
\name{reshape_forcis}
\alias{reshape_forcis}
\title{Reshape and simplify FORCIS data}
\usage{
reshape_forcis(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset, except for CPR North data.}
}
\value{
A \code{data.frame} reshaped in a long format.
}
\description{
A short description...
}
\examples{
## ADD EXAMPLE ----
}
