% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_columns}
\alias{select_columns}
\title{Select columns in FORCIS data}
\usage{
select_columns(data, cols = NULL)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{cols}{a \code{character} vector of column names to keep in addition to the
required ones (see \code{\link[=get_required_columns]{get_required_columns()}}) and to the taxa columns. Can
be \code{NULL} (default).}
}
\value{
A \code{data.frame}.
}
\description{
Selects columns in FORCIS data. Because FORCIS data contains more than 100
columns, this function can be used to lighten the \code{data.frame} to easily
handle it and to speed up some computations.
}
\examples{
# Attach the package ----
library("forcis")

# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_pump_sample.csv"), 
                         package = "forcis")

pump_data <- vroom::vroom(file_name, delim = ";", show_col_types = FALSE)

# Add 'data_type' column ----
pump_data$"data_type" <- "Pump"

# Dimensions of the data.frame ----
dim(pump_data)

# Select a taxonomy ----
pump_data <- select_taxonomy(pump_data, taxonomy = "OT")

# Dimensions of the data.frame ----
dim(pump_data)

# Select only required columns (and taxa) ----
pump_data <- select_columns(pump_data)

# Dimensions of the data.frame ----
dim(pump_data)
}
