% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zen_get_version_info}
\alias{zen_get_version_info}
\title{Print information of a specific version of a Zenodo repository}
\usage{
zen_get_version_info(record_id = zen_record_id(), version = NULL)
}
\arguments{
\item{record_id}{a \code{character} of length 1. The identifier of the repository
in the Zenodo database.}

\item{version}{a \code{character} of length 1. The label of the version. Use
\code{\link[=zen_list_versions]{zen_list_versions()}} to list available versions. If \code{NULL} (default) the
latest version is used.}
}
\value{
A \code{list} with all information about the version, including: \code{title},
\code{doi}, \code{publication_date}, \code{description}, \code{access_right}, \code{creators},
\code{keywords}, \code{version}, \code{resource_type}, \code{license}, and \code{files}.
}
\description{
Print information of a specific version of a Zenodo repository
}
\examples{
## Get information for the latest version of the FORCIS database ----

zen_get_version_info()
}
