% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zen_list_versions}
\alias{zen_list_versions}
\title{Retrieve all versions of a Zenodo repository}
\usage{
zen_list_versions(record_id = zen_record_id())
}
\arguments{
\item{record_id}{a \code{character} of length 1. The identifier of the repository
in the Zenodo database.}
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{publication_date}: the date of the release of the version
\item \code{version}: the label of the version
\item \code{access_right}: is the version open or restricted?
}
}
\description{
Retrieve all versions of a Zenodo repository
}
\examples{
## Versions of the FORCIS database ----

zen_list_versions()
}
