% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_abundances.R,
%   R/compute_concentration_bins_CPR_n.R, R/compute_concentrations.R,
%   R/compute_frequencies.R
\name{compute_abundances}
\alias{compute_abundances}
\alias{compute_concentration_bins_CPR_n}
\alias{computations}
\alias{compute_concentrations}
\alias{compute_frequencies}
\title{Compute count conversions}
\usage{
compute_abundances(data, aggregate = TRUE)

compute_concentration_bins_CPR_n(data)

compute_concentrations(data, aggregate = TRUE)

compute_frequencies(data, aggregate = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{aggregate}{a \code{logical} of length 1. If \code{FALSE} counts will be derived
for each subsample. If \code{TRUE} (default) subsample counts will be
aggregated by \code{sample_id}.}
}
\value{
A \code{data.frame}.
}
\description{
\strong{ADD DESCRIPTION}
}
\details{
\itemize{
\item \code{compute_concentrations()} \strong{ADD A FEW WORDS}
\item \code{compute_frequencies()} \strong{ADD A FEW WORDS}
\item \code{compute_abundances()} \strong{ADD A FEW WORDS}
}
}
