% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{compute_concentrations}
\alias{compute_concentrations}
\title{Compute concentrations from net data}
\usage{
compute_concentrations(df, aggregate = TRUE)
}
\arguments{
\item{df}{forcis net dataset}

\item{aggregate}{logical, if FALSE counts will be derived for each subsample.
If TRUE subsample counts will be aggregated by sample_id.}
}
\value{
dataframe
}
\description{
Compute concentrations from net data
}
