% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_sf.R
\name{data_to_sf}
\alias{data_to_sf}
\title{Convert a data frame into an sf object}
\usage{
data_to_sf(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset or the output of a
\verb{filter_*()} function.}
}
\value{
An \verb{sf POINTS} object.
}
\description{
This function can be used to convert a data frame into an sf object.
}
\examples{
## ADD EXAMPLE ----
}
