% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_bbox.R
\name{filter_by_bbox}
\alias{filter_by_bbox}
\title{Filter FORCIS data by a spatial bounding box}
\usage{
filter_by_bbox(data, bbox)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{bbox}{an object of class \code{bbox} (package \code{sf}) or a vector of four
\code{numeric} values defining a square bounding box. Values must follow this
order: minimum longitude (\code{xmin}), minimum latitude (\code{ymin}), maximum
longitude (\code{xmax}), and maximum latitude (\code{ymax}).
\strong{Important:} if a vector of numeric values is provided, coordinates must
be defined in the system WGS 84 (\code{epsg=4326}).}
}
\value{
A \code{data.frame} containing a subset of \code{data} for the desired
bounding box.
}
\description{
Filters FORCIS data by a spatial bounding box.
}
\examples{
# Attach the package ----
library("forcis")

# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_pump_sample.csv"), 
                         package = "forcis")

pump_data <- vroom::vroom(file_name, delim = ";", show_col_types = FALSE)

# Add 'data_type' column ----
pump_data$"data_type" <- "Pump"

# Dimensions of the data.frame ----
dim(pump_data)

# Filter by oceans ----
pump_data_sub <- filter_by_bbox(pump_data, bbox = c(35, 0, 45, 30))

# Dimensions of the data.frame ----
dim(pump_data_sub)
}
