% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_month}
\alias{filter_by_month}
\title{Filter forcis data by month of sampling}
\usage{
filter_by_month(data, months)
}
\arguments{
\item{data}{forcis data in long format}

\item{months}{numeric vector of selected months}
}
\value{
A \code{data.frame}.
}
\description{
Filter forcis data by month of sampling
}
\examples{
## ADD EXAMPLE ----
}
