% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_ocean.R
\name{filter_by_ocean}
\alias{filter_by_ocean}
\title{Filter FORCIS data by ocean}
\usage{
filter_by_ocean(data, ocean)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}

\item{ocean}{a \code{character} vector of one or several ocean names. Use the
function \code{get_ocean_names()} to find the correct spelling.}
}
\value{
A \code{data.frame} containing a subset of \code{data}.
}
\description{
This function can be used to filter FORCIS data by one or several oceans.
}
\examples{
## ADD EXAMPLE ----
}
