% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_polygon.R
\name{filter_by_polygon}
\alias{filter_by_polygon}
\title{Filter FORCIS data by a spatial polygon}
\usage{
filter_by_polygon(data, polygon)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}

\item{polygon}{an \verb{sf POLYGON} object.}
}
\value{
A \code{data.frame} containing a subset of \code{data}.
}
\description{
This function can be used to filter FORCIS data by a spatial polygon.
}
\examples{
## ADD EXAMPLE ----
}
