% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_species.R
\name{filter_by_species}
\alias{filter_by_species}
\title{Filter FORCIS data by species}
\usage{
filter_by_species(data, species, rm_na = FALSE)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset, except for CPR North data.}

\item{species}{a \code{character} vector listing species of interest.}

\item{rm_na}{a \code{logical} value. If \code{FALSE}, keeps taxa with \code{NA} counts.}
}
\value{
A \code{data.frame} containing a subset of \code{data}.
}
\description{
A short description...
}
\examples{
## ADD EXAMPLE ----
}
