% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_species}
\alias{filter_by_species}
\title{Filter forcis data by species}
\usage{
filter_by_species(data, species)
}
\arguments{
\item{data}{forcis data in long format, except for CPR North data}

\item{species}{a character vector listing species of interest}
}
\value{
A \code{data.frame}
}
\description{
Filter forcis data by species
}
\examples{
## ADD EXAMPLE ----
}
