% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_species.R
\name{filter_by_species}
\alias{filter_by_species}
\title{Filter FORCIS data by species}
\usage{
filter_by_species(data, species, rm_na = FALSE)
}
\arguments{
\item{data}{a \code{data.frame}. Must be the output of the function
\code{reshape_data()}, i.e. a FORCIS dataset in long format.}

\item{species}{a \code{character} vector listing species of interest.}

\item{rm_na}{a \code{logical} value. If \code{FALSE} (default), keeps taxa with \code{NA}
counts.}
}
\value{
A \code{data.frame} containing a subset of \code{data}.
}
\description{
Filters FORCIS data by a species list.
}
\examples{
# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_net_sample.csv"), 
                         package = "forcis")

net_data <- read.table(file_name, dec = ".", sep = ";")

# Add 'data_type' column ----
net_data$"data_type" <- "Net"

# Select a taxonomy ----
net_data <- select_taxonomy(net_data, taxonomy = "VT")

# Select only required columns (and taxa) ----
net_data <- select_columns(net_data)

# Dimensions of the data.frame ----
dim(net_data)

# Convert data into long format ----
net_data <- reshape_data(net_data)

# Dimensions of the data.frame ----
dim(net_data)

# Total number of species ----
length(unique(net_data$"taxa"))

# Select records for three species ----
net_data_sub <- filter_by_species(data    = net_data, 
                                  species = c("g_inflata_VT", 
                                              "g_elongatus_VT", 
                                              "g_glutinata_VT"))

# Dimensions of the data.frame ----
dim(net_data_sub)

# Total number of species ----
length(unique(net_data_sub$"taxa"))
}
