% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_year.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter FORCIS data by year of sampling}
\usage{
filter_by_year(data, years)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{years}{a \code{numeric} containing one or several years.}
}
\value{
A \code{data.frame} containing a subset of \code{data} for the desired years.
}
\description{
Filters FORCIS data by year of sampling.
}
\examples{
# Attach the package ----
library("forcis")

# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_pump_sample.csv"), 
                         package = "forcis")

pump_data <- vroom::vroom(file_name, delim = ";", show_col_types = FALSE)

# Add 'data_type' column ----
pump_data$"data_type" <- "Pump"

# Dimensions of the data.frame ----
dim(pump_data)

# Filter by years ----
pump_data_sub <- filter_by_year(pump_data, years = 1985)

# Dimensions of the data.frame ----
dim(pump_data_sub)
}
