% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter forcis data by year of sampling}
\usage{
filter_by_year(data, years)
}
\arguments{
\item{data}{forcis data}

\item{years}{numeric vector of selected years}
}
\value{
A \code{data.frame}.
}
\description{
Filter forcis data by year of sampling
}
\examples{
## ADD EXAMPLE ----
}
