% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter FORCIS data by year of sampling}
\usage{
filter_by_year(data, years)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}

\item{years}{a \code{numeric} containing one or several years.}
}
\value{
A \code{data.frame} containing a subset of \code{data} for the desired years.
}
\description{
This function can be used to filter FORCIS data by year of sampling.
}
\examples{
## ADD EXAMPLE ----
}
