% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcis-package.R
\docType{package}
\name{forcis-package}
\alias{forcis}
\alias{forcis-package}
\title{forcis: A Client to Access the FORCIS Database}
\description{
Provides an interface to the FORCIS database on global foraminifera distribution. This package allows to download and to handle FORCIS data. It is part of the FRB-CESAB working group FORCIS. <https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/forcis/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frbcesab.github.io/forcis}
  \item Report bugs at \url{https://github.com/FRBCesab/forcis/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mattia Greco \email{mgreco@iopan.pl} (\href{https://orcid.org/0000-0003-2416-6235}{ORCID})
  \item Xavier Giraud \email{giraud@cerege.fr} (\href{https://orcid.org/0000-0001-5067-8176}{ORCID})
  \item Sonia Chaabane \email{sonia.chaabane@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-4653-8610}{ORCID})
}

}
\keyword{internal}
