% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_basemap.R
\name{geom_basemap}
\alias{geom_basemap}
\title{Add a World basemap to a ggplot object}
\usage{
geom_basemap()
}
\value{
A \code{ggplot} object.
}
\description{
This function creates a World base map that can be added to a ggplot object.
Spatial layers come from the Natural Earth project
(\url{https://www.naturalearthdata.com/}).
}
\examples{
## ADD EXAMPLE ----
}
