% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_versions.R
\name{get_available_versions}
\alias{get_available_versions}
\title{Get available versions of the FORCIS database}
\usage{
get_available_versions()
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{publication_date}: the date of the release of the version
\item \code{version}: the label of the version
\item \code{access_right}: is the version open or restricted?
}
}
\description{
Gets all available versions of the FORCIS database by querying the Zenodo API
(\url{https://developers.zenodo.org}).
}
\examples{
## Versions of the FORCIS database ----

get_available_versions()
}
