% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_version.R
\name{get_current_version}
\alias{get_current_version}
\title{Get the version of the FORCIS database currently used}
\usage{
get_current_version()
}
\value{
A \code{character} of length 1, i.e. the label of the version in use.
}
\description{
Returns the version of the FORCIS database currently used in the project.
This function will read the content of the hidden file \code{.forcis} created by
the function \code{download_forcis_db()}. This file keeps track of the latest
version of the database used for a dedicated project. For more information,
please read the vignette available at
\url{https://frbcesab.github.io/forcis/articles/database-versions.html}.
}
\examples{
\dontrun{
# Folder in which the database will be saved ----
path_to_save_db <- "data"

# Download the database ----
download_forcis_db(path = path_to_save_db, version = NULL)

# Get the version of the database ----
get_current_version()
}
}
