% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-funs.R
\name{get_forcis_db}
\alias{get_forcis_db}
\title{Download the last version of the FORCIS database}
\usage{
get_forcis_db(path = ".")
}
\arguments{
\item{path}{a character. The folder in which the FORCIS database will be
save}
}
\value{
No return value.
}
\description{
...
}
\examples{
## Download the FORCIS database ----
# forcis::get_forcis_db()
}
