% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_forcis_db}
\alias{get_forcis_db}
\title{Download the entire FORCIS database}
\usage{
get_forcis_db(path = ".", version = NULL, overwrite = FALSE, timeout = 60)
}
\arguments{
\item{path}{a \code{character} of length 1. The folder in which the FORCIS
database will be saved. Note that a subdirectory will be created in the
version number.}

\item{version}{a \code{character} of length 1. The version number of the
FORCIS database. Default is the latest version.}

\item{overwrite}{a \code{logical}. If \code{TRUE} it will override the downloaded
files of the FORCIS database. Default is \code{FALSE}.}

\item{timeout}{a \code{integer}. The timeout for downloading files from the
FORCIS database. Default is \code{60}.}
}
\value{
No return value. The FORCIS files will be saved in the \code{path} folder.
}
\description{
Downloads the entire FORCIS database as a collection of five \code{csv} files from
Zenodo (\url{https://zenodo.org/record/7936568}).
}
\examples{
\dontrun{
# Folder in which the database will be saved ----
path_to_save_db <- "data"

# Download the database ----
forcis::get_forcis_db(path = path_to_save_db)

# Check the content of the folder ----
list.files(path_to_save_db, recursive = TRUE)
}
}
