% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-funs.R
\name{get_plankton_nets_data}
\alias{get_plankton_nets_data}
\title{Import plankton nets data}
\usage{
get_plankton_nets_data(path, version = forcis_db_version())
}
\arguments{
\item{path}{a character. The folder in which the FORCIS database will be
save}

\item{version}{a character. The version number of the database.}
}
\value{
A \code{data.frame} with x rows and y columns. See \strong{Data Paper URL} for
further information.
}
\description{
This function reads the csv file \code{planktonnet_May2022.csv} stored in the
folder \code{path}. If this file does not exist it will be downloaded from
\url{Repository}.
}
