% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_names.R
\name{get_species_names}
\alias{get_species_names}
\title{Get species names from column names}
\usage{
get_species_names(data)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}
}
\value{
A \code{data.frame}.
}
\description{
Gets species names from column names. This function is just an utility to
easily retrieve taxon names.
}
\examples{
\dontrun{
# Folder in which the database is stored ----
path_to_db <- "data"

# Download and read the plankton nets data ----
plankton_nets_data <- read_plankton_nets_data(path_to_db)

# Select a taxonomy ----
plankton_nets_data <- select_taxonomy(plankton_nets_data, taxonomy = "OT")

# Retrieve taxon names ----
get_species_names(nets)
}
}
