% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.R
\name{get_species_names}
\alias{get_species_names}
\title{Get species names from column names}
\usage{
get_species_names(data)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{get_*_data()} functions.}
}
\value{
A \code{data.frame}.
}
\description{
Gets species names from column names. This function is just an utility to
easily retrieve taxon names.
}
\examples{
\dontrun{
# Folder in which the database is stored ----
path_to_db <- "data"

# Download and read the plankton nets data ----
nets <- forcis::get_plankton_nets_data(path_to_db)

# Select a taxonomy ----
nets <- forcis::select_taxonomy(nets, taxonomy = "OT")

# Retrieve taxon names ----
forcis::get_species_names(nets)
}
}
