% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_data.R
\name{ggmap_data}
\alias{ggmap_data}
\title{Plot a distribution map of FORCIS data}
\usage{
ggmap_data(data, col = "red", ...)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset or the output of a
\verb{filter_*()} function.}

\item{col}{a \code{character} of length 1. The color of data on the map.}

\item{...}{other graphical parameters passed on to \code{geom_sf()}.}
}
\value{
A \code{ggplot} object.
}
\description{
This function can be used to map FORCIS data.
}
\examples{
## ADD EXAMPLE ----
}
