% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_depth.R
\name{plot_record_by_depth}
\alias{plot_record_by_depth}
\title{Plot sample records by depth of collection}
\usage{
plot_record_by_depth(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a barplot of FORCIS sample records by depth.
}
\examples{
## ADD EXAMPLE ----
}
