% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_depth.R
\name{plot_record_by_depth}
\alias{plot_record_by_depth}
\title{Plot sample record by depth of collection}
\usage{
plot_record_by_depth(data)
}
\arguments{
\item{data}{forcis data}
}
\value{
barplot with sample counts by depth interval
}
\description{
Plot sample record by depth of collection
}
