% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_month.R
\name{plot_record_by_month}
\alias{plot_record_by_month}
\title{Plot sample records by month}
\usage{
plot_record_by_month(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a barplot of FORCIS sample records by month.
}
\examples{
## ADD EXAMPLE ----
}
