% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_month.R
\name{plot_record_by_month}
\alias{plot_record_by_month}
\title{Plot sample record by month}
\usage{
plot_record_by_month(data)
}
\arguments{
\item{data}{forcis data}
}
\value{
barplot with sample counts by month
}
\description{
Plot sample record by month
}
