% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_season.R
\name{plot_record_by_season}
\alias{plot_record_by_season}
\title{Plot sample record by season}
\usage{
plot_record_by_season(data)
}
\arguments{
\item{data}{forcis data}
}
\value{
barplot with sample counts by season
}
\description{
Plot sample record by season
}
