% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_year.R
\name{plot_record_by_year}
\alias{plot_record_by_year}
\title{Plot sample records by year}
\usage{
plot_record_by_year(data)
}
\arguments{
\item{data}{a \code{data.frame}, i.e. a FORCIS dataset.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a barplot of FORCIS sample records by year.
}
\examples{
## ADD EXAMPLE ----
}
