% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_record_by_year.R
\name{plot_record_by_year}
\alias{plot_record_by_year}
\title{Plot sample record by year}
\usage{
plot_record_by_year(data)
}
\arguments{
\item{data}{forcis data}
}
\value{
barplot with sample counts by year
}
\description{
Plot sample record by year
}
