% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{reshape_forcis}
\alias{reshape_forcis}
\title{Reshape and simplify forcis data}
\usage{
reshape_forcis(data)
}
\arguments{
\item{data}{forcis data}
}
\value{
A \code{data.frame}
}
\description{
Reshape and simplify forcis data
}
