% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.R
\name{select_columns}
\alias{select_columns}
\title{Select columns in a data.frame}
\usage{
select_columns(data, cols = NULL)
}
\arguments{
\item{data}{a data.frame. One obtained by \verb{get_*_data()} functions.}

\item{cols}{a character vector of columns names to keep in addition to the
required ones.}
}
\value{
A \code{data.frame}
}
\description{
Select columns in a data.frame
}
