% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.R
\name{select_taxonomy}
\alias{select_taxonomy}
\title{Select a taxonomy in FORCIS data}
\usage{
select_taxonomy(data, taxonomy)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{get_*_data()} functions.}

\item{taxonomy}{a \code{character} of length 1. One among \code{"LT"}, \code{"VT"}, \code{"OT"}.}
}
\value{
A \code{data.frame}.
}
\description{
\strong{ADD DESCRIPTION}
}
\examples{
## __ADD EXAMPLE__
}
