% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selects.R
\name{select_taxonomy}
\alias{select_taxonomy}
\title{Select a taxonomy in FORCIS data}
\usage{
select_taxonomy(data, taxonomy)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{taxonomy}{a \code{character} of length 1. One among \code{"LT"}, \code{"VT"}, \code{"OT"}.}
}
\value{
A \code{data.frame}.
}
\description{
Selects a taxonomy in FORCIS data. FORCIS database provides three different
taxonomies: \code{"LT"} (lumped taxonomy), \code{"VT"} (validated taxonomy) and \code{"OT"}
(original taxonomy). See \url{https://doi.org/10.1038/s41597-023-02264-2} for
further information.
}
\examples{
\dontrun{
# Folder in which the database is stored ----
path_to_db <- "data"

# Download and read the plankton nets data ----
nets <- forcis::read_plankton_nets_data(path_to_db)

# Select a taxonomy ----
nets <- forcis::select_taxonomy(nets, taxonomy = "OT")
}
}
