% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_taxonomy.R
\name{select_taxonomy}
\alias{select_taxonomy}
\title{Select a taxonomy in FORCIS data}
\usage{
select_taxonomy(data, taxonomy)
}
\arguments{
\item{data}{a \code{data.frame}. One obtained by \verb{read_*_data()} functions.}

\item{taxonomy}{a \code{character} of length 1. One among \code{"LT"}, \code{"VT"}, \code{"OT"}.}
}
\value{
A \code{data.frame}.
}
\description{
Selects a taxonomy in FORCIS data. FORCIS database provides three different
taxonomies: \code{"LT"} (lumped taxonomy), \code{"VT"} (validated taxonomy) and \code{"OT"}
(original taxonomy). See \url{https://doi.org/10.1038/s41597-023-02264-2} for
further information.
}
\examples{
# Attach the package ----
library("forcis")

# Import example dataset ----
file_name <- system.file(file.path("extdata", "FORCIS_pump_sample.csv"), 
                         package = "forcis")

pump_data <- vroom::vroom(file_name, delim = ";", show_col_types = FALSE)

# Add 'data_type' column ----
pump_data$"data_type" <- "Pump"

# Dimensions of the data.frame ----
dim(pump_data)

# Select a taxonomy ----
pump_data <- select_taxonomy(pump_data, taxonomy = "OT")

# Dimensions of the data.frame ----
dim(pump_data)
}
